# FIX - size_t to int or int32_t size_t가 사용된 부분들을 확인하여 다른 자료형으로 교체하였습니다.

+++ src/plugins/chain_plugin/config/storage_type.hpp
@@ -19,7 +19,7 @@ using timestamp_t = uint64_t;
 constexpr auto TRANSACTION_ID_TYPE_SIZE = 32;
 constexpr auto ALPHANUMERIC_ID_TYPE_SIZE = 8;
 
-using block_height_type = size_t;
+using block_height_type = int32_t;
 using alphanumeric_type = std::string;
 using base58_type = std::string;
 using base64_type = std::string;
